requirements.txt
broadcaster[redis]==0.2.0
jinja2==3.1.2
uvicorn==0.22.0
websockets==11.0.3
starlette==0.27.0
redis==4.6.0
python-dotenv
aiohttp==3.8.5